feat: Refine entity model and setup migrations
- Setup ContactDatabase to setup the database context using SQLite,
creating the database in the user's local app data folder.
- Add a helper method to the ContactDatabase class to apply pending
migrations. This will also create the database if it doesn't exist.
In future, backup of old database will be implemented,
and cases will be handled where the database is corrupted, or various
other issues that prevent a migration.
- The PhoneNumberCategory and EmailAddressCategory classes have been
altered to be actual entities in the database. They are pre-populated
with default values using "Model Managed Data".