feat: Add ConsoleUIService
Add new service for displaying a console user interface. The service is a
BackgroundService so that it will be run on setup, and last for the lifecycle of
the program. Program.cs has been adjusted to await for host.StartAsync and
host.StopAsync so that it will exit once there are no services running.