feat: Stop UI state system blocking main thread
Adjust the UI state machine system to not block the main thread. This is
achieved by making the ExecuteAsync method asyncronous, which it wasn't by
mistake before. The main loop runs while there is not a cancellation requested
by the host.

The Stop method has been adjusted to use
IHostApplicationLifetime to request application stop. This is called in the
ExitState class, which awaits for the stopping token to be cancelled, blocking
the current thread until the stoppingToken updates.

A significant issue
with the previous code was that Console.In.ReadLine was blocking the main
thread. This has been removed for now, and replaced with a temporary
Task.Delay call to simulate some time. A solution needs to be found for this
issue in the future