feat: Add rudimentary example of data requesting from front end
Inject ContactsLogic into TestState, to demonstrate requesting data from the
database on the front-end. ContactsLogic has an injected instance of
ContactsRepository, and queries lists of PhoneNumberCategories and
EmailAddressCategories. ContactsRepository provides these lists by interacting
with the ContactsDbContext directly. In ContactsLogic, the methods return string
lists containing the labels for both PhoneNumberCategories and
EmailAddressCategories. No testing has been added yet, as this is just an
example.